src/pyproject.toml
[tool.poetry]
name = "Analyst"
version = "0.1.0-alpha"
description = ""
authors = ["Ricardo Cataldi <rcataldi@microsoft.com>"]
readme = "README.md"
packages = [{include = "app", from = "."}]

[tool.poetry.dependencies]
python = "^3.12, <4.0"
fastapi = "^0.115.5"
semantic-kernel = "^1.26.1"
uvicorn = "*"
typing-extensions = "^4.9.0"  # Ensures compatibility with TypeAlias in Python <3.10
pandas = "^2.2.3"
pypdf = {extras = ["crypto"], version = "^5.4.0"}
numpy = "^2.2.4"
scipy = "^1.15.2"
seaborn = "^0.13.2"
azure-cosmos = "^4.9.0"
reportlab = "^4.3.1"
scikit-learn = "^1.6.1"


[tool.poetry.group.dev.dependencies]
faker = "*"
python-dotenv = "*"
poetry-plugin-export = "*"
jupyterlab = "^4.3.5"


[tool.poetry.group.test.dependencies]
pytest = "*"
pytest-asyncio = "*"
rouge = "*"


[tool.poetry.group.lint.dependencies]
pylint = "*"
black = {extras = ["jupyter"], version = "^25.1.0"}
isort = "*"


[tool.poetry.group.docs.dependencies]
mkdocs = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q -s"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "async: marks tests as async (select with '-m async')",
    "slow: marks tests as slow (select with '-m slow')"
]

[tool.pylint]
fail-under = 8.5
max-line-length = 100
max-complexity = 10
good-names = ["i", "j", "k", "ex", "Run", "_"]
disable = [
    "C0114", "C0115", "C0116", "C0301",  # Docstrings & Line Length Warnings
    "W3101",  # `confusing-with-statement` (False Positive)
    "E0401"  # `import-error` (Fixes false positives on local imports)
]
load-plugins = [
    "pylint.extensions.mccabe",
]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']  # Explicit Python 3.12 support
include = '\.pyi?$'
exclude = '\.git|\.hg|\.mypy_cache|\.tox|\.venv|_build|buck-out|build|dist'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 100
